+++ js/utils/shuffle-array.js
@@ -1,0 +1,15 @@
+// shuffle array
+
+const shuffleArray = (arr) => {
+  let j;
+  let temp;
+  for(let i = arr.length - 1; i > 0; i--) {
+    j = Math.floor(Math.random() * (i + 1));
+    temp = arr[j];
+    arr[j] = arr[i];
+    arr[i] = temp;
+  }
+  return arr;
+};
+
+export {shuffleArray};
